Add support for variables in value bindings
The syntax is `val variableName = value; nextExpression`.
The variable is not assignable again - it is always immutable.
Variable may shadow property names and variables defined ahead of it.
The variable may be accessed only in the `nextExpression`.

This version does not support automatic unwrapping Tasks.
We wanted to switch to using await in DotVVM 3 anyway, right?

see #776